+++ OceansApp/OceansApp.Models/ViewModels/Consultants/CreateUpdateConsultantVM.cs
@@ -32,7 +32,7 @@ namespace OceansApp.Models.ViewModels.Consultants
         public DateTime? StartDate { get; set; }
         public List<CreateUpdateConsultantsAndPositionsVM>? Positions { get; set; }
         public bool IsEligibleForPaidTimeOff { get; set; }
-        public int? AnnualPaidTimeOffDays { get; set; }
+        public decimal? AnnualPaidTimeOffDays { get; set; }
         public int? InitialPtoBalance { get; set; }
     }
 }
